requirements.txt

numpy
python-dotenv
tqdm
PyPDF2
markdown
html2text
tiktoken
beautifulsoup4
